Moved Shape file location
